worker/requirements.txt
Flask==2.1.3
requests==2.22.0
pandas==1.4.3
scikit-learn==1.1.2
python-dotenv>=1.0.0
retry==0.9.2
--extra-index-url https://test.pypi.org/simple/
mictlanx==0.0.37
rory==0.0.12
